+++ hapi-fhir-jpaserver-model/src/main/java/ca/uhn/fhir/jpa/model/config/BaseSubscriptionSettings.java
@@ -50,6 +50,14 @@ public abstract class BaseSubscriptionSettings {
 	private boolean myAllowOnlyInMemorySubscriptions = false;
 
 	/**
+	 * If this is enabled (default is {@literal false}), changes to Subscription resource would be put on queue immediately.
+	 * Reducing delay between creation of the Subscription and Activation.
+	 *
+	 * @since 7.6.0
+	 */
+	private boolean mySubscriptionChangeQueuedImmediately = false;
+
+	/**
 	 * @since 7.6.0
 	 *
 	 * Regex To perform validation on the endpoint URL for Subscription of type RESTHOOK.
@@ -273,5 +281,25 @@ public abstract class BaseSubscriptionSettings {
 	 */
 	public boolean hasRestHookEndpointUrlValidationRegex() {
 		return isNotBlank(myRestHookEndpointUrlValidationRegex);
+	}
+
+	/**
+	 * If this is enabled (default is {@literal false}), changes to Subscription resource would be put on queue immediately.
+	 * Reducing delay between creation of the Subscription and Activation.
+	 *
+	 * @since 7.6.0
+	 */
+	public boolean isSubscriptionChangeQueuedImmediately() {
+		return mySubscriptionChangeQueuedImmediately;
+	}
+
+	/**
+	 * If this is enabled (default is {@literal false}), changes to Subscription resource would be put on queue immediately.
+	 * Reducing delay between creation of the Subscription and Activation.
+	 *
+	 * @since 7.6.0
+	 */
+	public void setSubscriptionChangeQueuedImmediately(boolean theSubscriptionChangeQueuedImmediately) {
+		mySubscriptionChangeQueuedImmediately = theSubscriptionChangeQueuedImmediately;
 	}
 }
